TRAIN-134 refactor: ScenarioRepository queryDsl -> jpql로 변경, 테스트 결과 수정

+++ src/main/java/com/aid/train/backend/domain/scenario/service/ScenarioService.java
@@ -106,7 +106,7 @@ public class ScenarioService {
      * 사용자가 생성한 시나리오를 삭제합니다.
      */
     public String deleteScenario(Long userId, Long scenarioId) {
-        long flag = scenarioRepository.deleteByAndUserId(userId, scenarioId);
+        int flag = scenarioRepository.deleteByAndUserId(userId, scenarioId);
         if(flag == 0) {
             throw new TrainException(SCENARIO_NOT_FOUND);
         }
